refactor: relocate shared navigation components
