+++ src/Components/TopNavbar.jsx
@@ -43,7 +43,7 @@ const TopNavbar = () => {
         setTask([]);
       }
     } catch (err) {
-      console.log('Error fetching data:', err);
+      console.error('Error fetching data:', err);
     }
   };
